Wrote the skeleton/architecture of the extension
Wrote all of the messaging logic for communication between the background,
content, and popup scripts. It has not yet been tested, but should be close
to good if it isn't working properly already. I also added a popup to make
scraping a little more user friendly. Last, I updated server.js to listen
for the correct request, but did not update how it handles the request.