Collecting profile urls is now complete.
Fixed the logic to collect profile urls and send them to the background.
I also added an alert in the content script to tell us when it's done
successfully. I also fixed up some of the scrape logic, however it's not
done yet. Currently, the page will redirect before scraping, but the scraping
isn't quite working.